[TASK] Adapt cHash handling for TYPO3 10
cHash handling ws adapted/disabled to fix generated links in TYPO3 9,
as this adaptions lead to invalid links, e.g. facet links, in TYPO3 10
and Site configuration support is available since EXT:solr 10 there is
no need for this adaptions.

This commit removes the implemented removal of the cHash parameter,
now the UriBuilder is used to build the link template.

Resolves: #2725